move queries to pkg
