+++ cli/query.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/jmpsec/osctrl/queries"
+	"github.com/jmpsec/osctrl/pkg/queries"
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 )
