+++ pkg/logging/db.go
@@ -10,7 +10,7 @@ import (
 	"gorm.io/gorm"
 
 	"github.com/jmpsec/osctrl/backend"
-	"github.com/jmpsec/osctrl/queries"
+	"github.com/jmpsec/osctrl/pkg/queries"
 	"github.com/jmpsec/osctrl/settings"
 	"github.com/jmpsec/osctrl/types"
 	"github.com/rs/zerolog/log"
