+++ pkg/queries/utils.go
